linalg/wasm: move the f32 MMM kernels and the dispatch tests into the wasm/ module tree
Splits the six f32 kernels by accumulator layout: wasm/mmm_f32_gemm.rs holds the
two that pack columns into lanes (4x4, 8x8), wasm/mmm_f32_gemv.rs the four that
pack rows and fix nr=1 (4x1, 8x1, 16x1, 32x1) — the same split the mmm_f32 and
mmv_f32 dispatch callbacks make. The dispatch assertions move to
wasm/dispatch_tests.rs. Pure relocation: the emitted kernel bodies are unchanged
under both +simd128 and +simd128,+relaxed-simd.
